+++ lib/unilitix_flutter.dart
@@ -1,1 +1,10 @@
-export 'unilitix.dart';
+// ignore: deprecated_member_use_from_same_package
+export 'unilitix.dart'
+    show
+        Unilitix,
+        UnilitixConfig,
+        UnilitixObserver,
+        UnilitixApp,
+        UnilitixWidget,
+        UnilitixMaterialApp,
+        UnilitixPrivate;